+++ lib/src/models/santali_calendar_day.dart
@@ -4,14 +4,18 @@ class SantaliCalendarDay {
   final int day;
   final DateTime date;
   final bool isToday;
-  final SantaliWeekDay weekDay;
   final bool isCurrentMonth;
+  final bool isPurnima;
+  final bool isAmavasya;
+  final SantaliWeekDay weekDay;
 
   const SantaliCalendarDay({
     required this.day,
     required this.date,
-    required this.isToday,
     required this.weekDay,
+    required this.isToday,
+    required this.isPurnima,
+    required this.isAmavasya,
     required this.isCurrentMonth,
   });
 }