+++ lib/src/models/santali_calendar_month.dart
@@ -1,17 +1,21 @@
 import 'package:santali_calendar/src/models/santali_calendar_day.dart';
+import 'package:santali_calendar/src/models/santali_month.dart';
 
-class SantaliCalendarMonth {
-  final String name;
-  final String english;
-  final DateTime startDate;
-  final DateTime endDate;
+class SantaliCalendarMonth extends SantaliMonth {
   final List<SantaliCalendarDay?> days;
 
   const SantaliCalendarMonth({
     required this.days,
-    required this.name,
-    required this.english,
-    required this.endDate,
-    required this.startDate,
+    required super.id,
+    required super.name,
+    required super.roman,
+    required super.index,
+    required super.endDate,
+    required super.startDate,
+    required super.totalDays,
+    required super.isLeapMonth,
+    required super.newMoonDate,
+    required super.fullMoonDate,
+    required super.displayEndDate,
   });
 }